middleware/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlmodel
psycopg2-binary
pytest           # per executar els tests dins el contenidor
httpx
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytz
pandas
ace_tools